Rip out T4 stuff; replace with dumb script
I like T4, but the T4BuildTasks don't quite work in the new .NET world:
the binary does not work in the new .NET environment, and apparently
that is needed for "dotnet publish" to work. I have requested
.NET-compatible binaries from the T4 people; no idea if/when that will
ever happen.

A possible replacement would be to use mono/T4, but it wouldn't be a
drop-in replacement (there is no equivalent to the T4BuildTasks package,
with build tasks and .targets file). Since what I was actually doing
with T4 was really simple, I just wrote a dumb PowerShell script to do
the job instead.